refactor(driver-support): move stage clearing into stage.rs
GC's stage teardown now lives next to the staging mechanism as
`hdriver_support::stage::clear_stage`; engine's gc_stage delegates to it
(engine already depends on driver-support — no new edge). dir_size and
the per-entry-lock teardown move with it. Tests follow.
