+++ crates/driver-support/Cargo.toml
@@ -18,6 +18,7 @@ serde_json = "1.0"
 tar = "0.4"
 glob = "0.3"
 walkdir = "2"
+tracing = "0.1"
 xxhash-rust = { version = "0.8.15", features = ["xxh3", "std"] }
 tokio = { version = "1.52.3", features = ["rt", "io-util", "fs", "process"] }
