+++ crates/driver-support/src/stage.rs
@@ -234,6 +234,98 @@ pub fn stage_group_for(stage_root: &Path, key_prefix: &str) -> PathBuf {
     stage_root.join(sanitize_component(key_prefix))
 }
 
+/// Delete every staged entry under `stage_root`. Stage entries are a pure cache
+/// — a future consumer re-materializes whatever it needs — so GC just clears
+/// them all. Each `<group>/<hash>` entry is removed under its advisory lock so
+/// an in-flight materialization is never deleted mid-write; a contended entry
+/// is left for the next sweep. Best-effort: per-entry failures are logged and
+/// skipped. Returns `(entries_removed, bytes_freed)`.
+pub fn clear_stage(stage_root: &Path) -> (usize, u64) {
+    let groups = match std::fs::read_dir(stage_root) {
+        Ok(d) => d,
+        Err(e) if e.kind() == std::io::ErrorKind::NotFound => return (0, 0),
+        Err(e) => {
+            tracing::warn!(error = %e, path = %stage_root.display(), "clear_stage: read stage root");
+            return (0, 0);
+        }
+    };
+    let mut removed = 0usize;
+    let mut bytes = 0u64;
+    for group in groups.flatten() {
+        if !group.file_type().map(|t| t.is_dir()).unwrap_or(false) {
+            continue;
+        }
+        let gpath = group.path();
+        let entries = match std::fs::read_dir(&gpath) {
+            Ok(e) => e,
+            Err(e) => {
+                tracing::warn!(error = %e, path = %gpath.display(), "clear_stage: read group");
+                continue;
+            }
+        };
+        for ent in entries.flatten() {
+            // Only the `<hash>` dirs are entries; the `.ready`/`.lock` sidecars
+            // are reclaimed alongside their dir.
+            if !ent.file_type().map(|t| t.is_dir()).unwrap_or(false) {
+                continue;
+            }
+            let hash = ent.file_name().to_string_lossy().into_owned();
+            // Non-blocking: a held lock means a materializer is mid-write — skip
+            // and let the next sweep reclaim it.
+            let lock = FLock::new(gpath.join(format!("{hash}.lock")));
+            let _guard = match lock.try_lock() {
+                Ok(Some(g)) => g,
+                Ok(None) => continue,
+                Err(e) => {
+                    tracing::warn!(error = %format!("{e:#}"), %hash, "clear_stage: lock entry");
+                    continue;
+                }
+            };
+            let entry_dir = ent.path();
+            let freed = dir_size(&entry_dir);
+            if let Err(e) = hcore::fsutil::remove_dir_all(&entry_dir) {
+                tracing::warn!(error = %e, path = %entry_dir.display(), "clear_stage: remove entry");
+                continue;
+            }
+            // Drop the readiness witness; the `.lock` is removed when the guard
+            // releases (write-unlock deletes its lock file).
+            let ready = gpath.join(format!("{hash}.ready"));
+            match std::fs::remove_file(&ready) {
+                Ok(()) => {}
+                Err(e) if e.kind() == std::io::ErrorKind::NotFound => {}
+                Err(e) => {
+                    tracing::warn!(error = %e, path = %ready.display(), "clear_stage: remove ready")
+                }
+            }
+            removed += 1;
+            bytes = bytes.saturating_add(freed);
+        }
+    }
+    (removed, bytes)
+}
+
+/// Best-effort recursive byte size of `dir` (regular files only), for GC byte
+/// accounting. Errors are swallowed — a miscount never blocks a reclaim.
+fn dir_size(dir: &Path) -> u64 {
+    let mut total = 0u64;
+    let Ok(entries) = std::fs::read_dir(dir) else {
+        return total;
+    };
+    for entry in entries.flatten() {
+        let Ok(ft) = entry.file_type() else {
+            continue;
+        };
+        if ft.is_dir() {
+            total = total.saturating_add(dir_size(&entry.path()));
+        } else if ft.is_file()
+            && let Ok(md) = entry.metadata()
+        {
+            total = total.saturating_add(md.len());
+        }
+    }
+    total
+}
+
 #[cfg(all(test, unix))]
 mod tests {
     use super::*;
@@ -447,6 +539,56 @@ mod tests {
             body.contains(link.join("x.txt").to_str().unwrap()),
             "list must record the linked path, got: {body}"
         );
+    }
+
+    /// Create a stage entry `<root>/<group>/<hash>/blob` plus its `.ready`
+    /// witness; returns the entry dir.
+    fn make_stage_entry(root: &Path, group: &str, hash: &str) -> PathBuf {
+        let gdir = root.join(group);
+        let entry = gdir.join(hash);
+        std::fs::create_dir_all(&entry).expect("mkdir");
+        std::fs::write(entry.join("blob"), b"staged-bytes").expect("blob");
+        std::fs::write(gdir.join(format!("{hash}.ready")), b"").expect("ready");
+        entry
+    }
+
+    #[test]
+    fn clear_stage_deletes_all_entries() {
+        let tmp = tempfile::tempdir().expect("tempdir");
+        let root = tmp.path().join("stage");
+        let a = make_stage_entry(&root, "g1", "aaaa");
+        let b = make_stage_entry(&root, "g2", "bbbb");
+
+        let (removed, bytes) = clear_stage(&root);
+
+        assert_eq!(removed, 2, "every entry cleared");
+        assert_eq!(bytes, 2 * b"staged-bytes".len() as u64);
+        assert!(!a.exists());
+        assert!(!b.exists());
+        assert!(!root.join("g1/aaaa.ready").exists());
+        assert!(!root.join("g1/aaaa.lock").exists());
+    }
+
+    #[test]
+    fn clear_stage_missing_root_is_noop() {
+        let tmp = tempfile::tempdir().expect("tempdir");
+        assert_eq!(clear_stage(&tmp.path().join("nope")), (0, 0));
+    }
+
+    #[tokio::test]
+    async fn clear_stage_skips_locked_entry() {
+        let tmp = tempfile::tempdir().expect("tempdir");
+        let root = tmp.path().join("stage");
+        let busy = make_stage_entry(&root, "g", "held");
+
+        // Hold the entry's lock: a materializer is mid-write → skip it.
+        let lock = FLock::new(root.join("g/held.lock"));
+        let _held = lock.lock(&ct()).await.expect("hold lock");
+
+        let (removed, _bytes) = clear_stage(&root);
+
+        assert_eq!(removed, 0, "locked entry not reclaimed");
+        assert!(busy.exists(), "in-flight stage entry survives");
     }
 
     #[tokio::test]
